developed simple login system

+++ src/main/java/com/github/GabsOda/creditAnalysisAPI/service/ClientService.java
@@ -1,10 +1,13 @@
 package com.github.GabsOda.creditAnalysisAPI.service;
 
 import com.github.GabsOda.creditAnalysisAPI.dto.request.ClientDTO;
+import com.github.GabsOda.creditAnalysisAPI.dto.request.ClientLoginDTO;
+import com.github.GabsOda.creditAnalysisAPI.dto.request.ClientLogoutDTO;
 import com.github.GabsOda.creditAnalysisAPI.dto.request.ClientResponseDTO;
 import com.github.GabsOda.creditAnalysisAPI.dto.response.MessageResponseDTO;
 import com.github.GabsOda.creditAnalysisAPI.entity.Client;
 import com.github.GabsOda.creditAnalysisAPI.exception.ClientNotFoundException;
+import com.github.GabsOda.creditAnalysisAPI.exception.ClientNotLogged;
 import com.github.GabsOda.creditAnalysisAPI.mapper.ClientMapper;
 import com.github.GabsOda.creditAnalysisAPI.repository.ClientRepository;
 import lombok.AllArgsConstructor;
@@ -24,9 +27,49 @@ public class ClientService {
 
     public MessageResponseDTO createClient(ClientDTO clientDTO) {
         Client clientToSave = clientMapper.toModel(clientDTO);
+        clientToSave.setLoggedIn(false);
+
         Client savedClient = clientRepository.save(clientToSave);
 
         return createMessageResponse(savedClient.getId(), "Created Client with ID: ");
+    }
+
+    public MessageResponseDTO loginClient(ClientLoginDTO clientLoginDTO) throws ClientNotFoundException {
+        List<Client> client = clientRepository.findByEmail(clientLoginDTO.getEmail());
+
+        if(!client.isEmpty()){
+            Client clientLogged = client.get(0);
+            clientLogged.setLoggedIn(true);
+
+            Client updatedClient = clientRepository.save(clientLogged);
+
+            return createMessageResponse(updatedClient.getEmail(), "Logged client with email: ");
+        } else {
+            throw new ClientNotFoundException("Not found client with email: ", clientLoginDTO.getEmail());
+        }
+
+    }
+
+    public MessageResponseDTO logoutClient(ClientLogoutDTO clientLogoutDTO) throws ClientNotFoundException, ClientNotLogged {
+        List<Client> client = clientRepository.findByEmail(clientLogoutDTO.getEmail());
+
+        if (!client.isEmpty()) {
+            Client clientLoggedOut = client.get(0);
+
+            if (verifyIfClientIsLogged(clientLoggedOut.getId())){
+                clientLoggedOut.setLoggedIn(false);
+
+                Client updatedClient = clientRepository.save(clientLoggedOut);
+
+                return createMessageResponse(updatedClient.getEmail(), "Logged out client with email: ");
+            } else {
+                throw new ClientNotLogged("Not logged client with email: " + clientLoggedOut.getEmail());
+            }
+
+        } else {
+            throw new ClientNotFoundException("Not found client with email: ", clientLogoutDTO.getEmail());
+        }
+
     }
 
     public List<ClientResponseDTO> listAll() {
@@ -35,35 +78,60 @@ public class ClientService {
                 .collect(Collectors.toList());
     }
 
-    public ClientDTO findById(Long id) throws ClientNotFoundException{
-        return clientMapper.toDto(verifyIfClientExists(id));
+    public ClientDTO findById(Long id) throws ClientNotFoundException, ClientNotLogged {
+        if (verifyIfClientIsLogged(id)) {
+            return clientMapper.toDto(verifyIfClientExists(id));
+        } else {
+            throw new ClientNotLogged("Not logged client with ID: " + id);
+        }
     }
 
-    public MessageResponseDTO updateById(Long id, ClientDTO clientDTO) throws ClientNotFoundException {
+    public MessageResponseDTO updateById(Long id, ClientDTO clientDTO) throws ClientNotFoundException, ClientNotLogged {
         verifyIfClientExists(id);
 
-        Client clientToUpdate = clientMapper.toModel(clientDTO);
-        Client updatedClient = clientRepository.save(clientToUpdate);
+        if (verifyIfClientIsLogged(id)) {
+            Client clientToUpdate = clientMapper.toModel(clientDTO);
+            Client updatedClient = clientRepository.save(clientToUpdate);
 
-        return createMessageResponse(updatedClient.getId(), "Updated Client with ID: ");
+            return createMessageResponse(updatedClient.getId(), "Updated Client with ID: ");
+        } else {
+            throw new ClientNotLogged("Not logged client with ID: " + id);
+        }
+
     }
 
-    public MessageResponseDTO deleteById(Long id) throws ClientNotFoundException {
+    public MessageResponseDTO deleteById(Long id) throws ClientNotFoundException, ClientNotLogged {
         verifyIfClientExists(id);
 
-        clientRepository.deleteById(id);
+        if (verifyIfClientIsLogged(id)) {
+            clientRepository.deleteById(id);
 
-        return createMessageResponse(id, "Deleted client with ID: ");
+            return createMessageResponse(id, "Deleted client with ID: ");
+        } else {
+            throw new ClientNotLogged("Not logged client with ID: " + id);
+        }
     }
 
-    public Client verifyIfClientExists(Long id) throws ClientNotFoundException{
+    public Client verifyIfClientExists(Long id) throws ClientNotFoundException {
         return clientRepository.findById(id)
                 .orElseThrow(() -> new ClientNotFoundException("Not found client with Id: ", id));
+    }
+
+    public boolean verifyIfClientIsLogged(Long id) {
+        return clientRepository.findById(id)
+                .orElseThrow(() -> new ClientNotFoundException("Not found client with Id: ", id))
+                .getLoggedIn();
     }
 
     private MessageResponseDTO createMessageResponse(Long id, String message) {
         return MessageResponseDTO.builder()
                 .message(message + id)
+                .build();
+    }
+
+    private MessageResponseDTO createMessageResponse(String email, String message) {
+        return MessageResponseDTO.builder()
+                .message(message + email)
                 .build();
     }
